requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
airflow-dbt # jika ingin menggunakan operator
dbt-postgres # jika ingin menggunakan dbt run melalui bash operator